refactor(vm): isolate bytecode executor
